[DOCS] Fixes in `emma-quickstart` archetype.
- Fixed CLI file names and paths.
- Use `readText` to read in kMeans input.
- Improved WordCount algorithm tests.
- Map k-means result before writing.
  Resolves a problem with missing `Encoder[Vector[Double]]` for Spark.
- Shorter test nesting hierarchy.
- Fixed package structure.
  - Algorithms are located in `org.example.algorithms`.
  - CLI objects are located in `org.example.cli`.
