+++ emma-quickstart/src/main/resources/META-INF/maven/archetype-metadata.xml
@@ -25,12 +25,6 @@
             <include>**/*.scala</include>
           </includes>
         </fileSet>
-        <fileSet encoding="UTF-8">
-          <directory>src/main/resources</directory>
-          <includes>
-            <include>**/*.csv</include>
-          </includes>
-        </fileSet>
         <fileSet filtered="true" packaged="true" encoding="UTF-8">
           <directory>src/test/scala</directory>
           <includes>
@@ -40,6 +34,7 @@
         <fileSet encoding="UTF-8">
           <directory>src/test/resources</directory>
           <includes>
+            <include>**/*.txt</include>
             <include>**/*.tsv</include>
           </includes>
         </fileSet>
@@ -59,12 +54,6 @@
             <include>**/*.scala</include>
           </includes>
         </fileSet>
-        <fileSet filtered="true" encoding="UTF-8">
-          <directory></directory>
-          <includes>
-            <include>dependency-reduced-pom.xml</include>
-          </includes>
-        </fileSet>
       </fileSets>
     </module>
     <module id="${rootArtifactId}-spark" dir="__rootArtifactId__-spark" name="${rootArtifactId}-spark">
@@ -79,12 +68,6 @@
           <directory>src/test/scala</directory>
           <includes>
             <include>**/*.scala</include>
-          </includes>
-        </fileSet>
-        <fileSet filtered="true" encoding="UTF-8">
-          <directory></directory>
-          <includes>
-            <include>dependency-reduced-pom.xml</include>
           </includes>
         </fileSet>
       </fileSets>
