--- emma-quickstart/src/main/resources/archetype-resources/__rootArtifactId__-library/src/main/resources/graphs/dwg.csv
@@ -1,3 +1,0 @@
-1	2	John
-1	3	Jake
-3	5	Jillian
