--- emma-quickstart/src/main/resources/archetype-resources/__rootArtifactId__-library/src/main/scala/ml/model.scala
@@ -1,14 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-package ${package}
-package ml
-
-import breeze.linalg.{Vector => Vec}
-import org.emmalanguage.api._
-
-/** Machine learning model objects. */
-object model {
-  case class Point[ID](@emma.pk id: ID, vector: Vec[Double])
-  case class LVector[L](label: L, vector: Vec[Double])
-}
