--- emma-quickstart/src/main/resources/archetype-resources/__rootArtifactId__-library/src/main/scala/text/WordCount.scala
@@ -1,25 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-package ${package}
-package text
-
-import org.emmalanguage.api._
-
-@emma.lib
-object WordCount {
-
-  def apply(docs: DataBag[String]): DataBag[(String, Long)] = {
-    val words = for {
-      line <- docs
-      word <- DataBag[String](line.toLowerCase.split("${symbol_escape}${symbol_escape}W+"))
-    } yield word
-
-    // group the words by their identity and count the occurrence of each word
-    val counts = for {
-      group <- words.groupBy(identity)
-    } yield (group.key, group.values.size)
-
-    counts
-  }
-}
