+++ emma-quickstart/src/main/resources/archetype-resources/__rootArtifactId__-library/src/test/resources/ml/clustering/kmeans/clusters.tsv
