+++ emma-quickstart/src/main/resources/archetype-resources/__rootArtifactId__-spark/pom.xml
@@ -80,7 +80,7 @@
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer" />
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>org.emmalanguage.cli.SparkExamplesRunner</mainClass>
+                                    <mainClass>${package}.cli.SparkRunner</mainClass>
                                 </transformer>
                             </transformers>
                             <artifactSet>
