+++ emma-quickstart/src/test/resources/projects/basic/archetype.properties
@@ -1,4 +1,4 @@
-#Tue Apr 11 13:04:44 EEST 2017
+#Thu Apr 13 02:50:25 EEST 2017
 package=it.pkg
 version=0.1-SNAPSHOT
 groupId=archetype.it
